Replace User Password with Internal Token
Update the application to continue to use Tokens that belong to the
User, but define them by their usecase; either External or Internal
Tokens. Internals Tokens are only used within the Application and their
life time is brief since they can be created and destroyed easily.
External Tokens however should not be destroy, since they are used by
Applications external to this application and often requires manual
registration by an User to enable.